% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{file_cd_current}
\alias{file_cd_current}
\title{In Files pane, navigate to folder of file currently open in editor}
\usage{
file_cd_current()
}
\description{
In Files pane, navigate to folder of file currently open in editor
}
