% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{navigate_to_active_file}
\alias{navigate_to_active_file}
\title{In Files pane, navigate to folder of file currently open in editor}
\usage{
navigate_to_active_file()
}
\description{
In Files pane, navigate to folder of file currently open in editor
}
