% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wrap_in_if}
\alias{wrap_in_if}
\title{Wrap selection in an if statement}
\usage{
wrap_in_if()
}
\description{
Wraps the current line in an \verb{if (...) \{ <text> \}} block and places the
cursor inside the \verb{if ()} parentheses. If text is selected, wraps all lines
that are at least partially included in the selection.
}
