% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{navigate_to_active_file}
\alias{navigate_to_active_file}
\title{Navigate to directory of open file}
\usage{
navigate_to_active_file()
}
\description{
Open the Files pane and navigate to the folder containing the file currently
open in the editor.
}
